# api-gateway/cmd/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Manusia API
  description: |
    Marketplace jasa umum — siapapun bisa minta tolong atau jadi penolong.
    Semua request melalui API Gateway di port **8080**.

    **Auth**: Gunakan tombol "Authorize" → masukkan token JWT dengan format `Bearer <token>`.
    Token didapat dari `POST /api/auth/login`.
  version: "1.0.0"
  contact:
    name: Nicko
servers:
  - url: http://localhost:8080
    description: Development (via API Gateway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authId:
          type: string
        name:
          type: string
        bio:
          type: string
        location:
          type: string
        age:
          type: integer
        experience:
          type: integer
        rating:
          type: number
          format: float
        totalReviews:
          type: integer
        pricePerDay:
          type: integer
        workStatus:
          type: string
          enum: [OPEN, CLOSED, BOOKED]
        specializations:
          type: array
          items:
            type: string
        avatar:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        gender:
          type: string
        birthPlace:
          type: string
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
        customerName:
          type: string
        title:
          type: string
        description:
          type: string
        budgetPerDay:
          type: integer
          example: 150000
        todoList:
          type: array
          items:
            type: string
        durationDays:
          type: integer
        city:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        category:
          type: string
          enum: [TASK, PROJECT, EVENT]
        status:
          type: string
          enum: [OPEN, CLOSED]
        createdAt:
          type: string
          format: date-time
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workerId:
          type: string
        workerName:
          type: string
        workerAvatar:
          type: string
        customerId:
          type: string
        customerName:
          type: string
        address:
          type: string
        city:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        bookingDate:
          type: string
          example: "2026-07-01"
        startTime:
          type: string
          example: "08:00"
        durationDays:
          type: integer
        paymentMethod:
          type: string
          example: CASH
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
        notes:
          type: string
        jobId:
          type: string
        jobTitle:
          type: string
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /api/workers:
    get:
      tags: [Users & Workers]
      summary: List semua worker/penolong
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List worker
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
...
  /api/jobs:
    post:
      tags: [Jobs]
      summary: Buat posting pekerjaan baru
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerName, title, description, budgetPerDay, durationDays, city, category]
              properties:
                customerName:
                  type: string
                  example: Andi Wijaya
                title:
                  type: string
                  example: Bantu pindahan rumah
                description:
                  type: string
                budgetPerDay:
                  type: integer
                  example: 200000
                todoList:
                  type: array
                  items:
                    type: string
                durationDays:
                  type: integer
                  example: 2
                city:
                  type: string
                  example: Jakarta
                latitude:
                  type: number
                longitude:
                  type: number
                category:
                  type: string
                  enum: [TASK, PROJECT, EVENT]
      responses:
        '201':
          description: Pekerjaan berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
...
  /api/bookings:
    post:
      tags: [Bookings]
      summary: Buat booking langsung ke worker
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workerId, customerName, address, city, bookingDate, startTime, durationDays, paymentMethod]
              properties:
                workerId:
                  type: string
                workerName:
                  type: string
                workerAvatar:
                  type: string
                customerName:
                  type: string
                address:
                  type: string
                city:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                bookingDate:
                  type: string
                  example: "2026-07-15"
                startTime:
                  type: string
                  example: "08:00"
                durationDays:
                  type: integer
                paymentMethod:
                  type: string
                  example: CASH
                notes:
                  type: string
      responses:
        '201':
          description: Booking berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
...